Move placement cursor creation to DiagramView
